standardized naming and pep8
