+++ s3_parq/publish_parq.py
@@ -3,7 +3,7 @@ import pandas as pd
 from collections import namedtuple
 
 
-class S3ParqPublish:
+class S3PublishParq:
 
     def __init__(self, dataset: str, dataframe: pd.DataFrame, bucket: str, key_prefix: str = '')->None:
         pass
